Deepen mutter_port backend stubs and fix review findings
Second pass over the backend port:

- Deepen the 17 partially-built files that still carried placeholder
  TODO blocks (monitor_manager(+private), renderer_native_private,
  stage_private, native/kms_{crtc,connector,plane}_private, backlight_
  ref_white(+private), backlight_sysfs(+private), backend, clipboard_
  session, color_calibration_session, input_mapper_private,
  input_capture(+session)) — extracting the real enums, structs,
  property tables, and constants from upstream Mutter and removing the
  stale placeholder comments.

- Address the automated review of the port:
  - monitor_config_utils: import alloc::string::String (was used in a
    Result<(), String> return but not in scope in no_std).
  - screen_cast_stream: drop the import of the non-existent
    MetaScreenCastFlag type (screen_cast exposes u32 consts, not a type).
  - logical_monitor_private: derive Copy/PartialEq on MetaLogicalMonitorId
    and implement dup/equal via the derives (equal previously always
    returned false).
  - eis, input_settings, native/kms_types_private: model bitmask flags as
    `pub type ... = u32` + consts instead of enums so they can be OR-ed.
  - idle_monitor_private, gles3_table: use Option<extern "C" fn> for the
    optional callback/GL function pointer instead of a usize sentinel.
  - cursor_renderer: calculate_rect now returns (x, y, w, h).
  - launcher: activate_vt takes the VT as u32 (matches logind ActivateVT).

Wired-in kernel `cargo check` remains green.

+++ src/mutter_port/backends/launcher.rs
@@ -25,7 +25,7 @@ impl MetaLauncher {
     }
 
     /// Activate a virtual terminal.
-    pub fn activate_vt(&self, _vt: i8) -> Result<(), String> {
+    pub fn activate_vt(&self, _vt: u32) -> Result<(), String> {
         // TODO: D-Bus call to ActiveSession (vt, u)
         Ok(())
     }
